requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
